deafult page is now home page + styling fixes
