+++ client/src/app/components/upload-audio.js
@@ -8,7 +8,7 @@ const UploadAudio = () => {
           Upload Audio:
         </h1>
 
-        <Button className="w-full rounded-[40px] cursor-pointer" sx={{textTransform: "none", color: "black", width: 350, height: 310, bgcolor: '#F1F7FF', transition: "transform 0.3s, box-shadow 0.3s", "&:hover": {transform: "scale(1.05)", boxShadow: '0px 2px 10px rgba(0, 0, 0, 0.1)',} }}>
+        <Button className="w-full rounded-[40px] cursor-pointer" sx={{borderRadius: "25px", textTransform: "none", color: "black", width: 350, height: 310, bgcolor: '#F1F7FF', transition: "transform 0.3s, box-shadow 0.3s", "&:hover": {transform: "scale(1.05)", boxShadow: '0px 2px 10px rgba(0, 0, 0, 0.1)',} }}>
           <Box className="flex flex-col items-center pt-1 px-10 bg-[#F1F7FF]">
             <img
               className="w-[230px] h-50 mb-8"
