+++ client/src/app/pages/use-the-tool/page.js
@@ -5,6 +5,7 @@ import UploadAudio from "@/app/components/upload-audio";
 import RecordAudio from "@/app/components/record-audio";
 import SubmitAudioHeading from "@/app/components/submit-audio-heading";
 import Navbar from "@/app/components/navbar";
+import Link from "next/link";
 
 export default function Action() {
   return (
@@ -13,7 +14,7 @@ export default function Action() {
     
         
         <Stack width= "100%" direction ="column" alignItems="center" spacing={5} sx={{ justifyContent: "center", mt: 6 }} >
-          <SubmitAudioHeading></SubmitAudioHeading>
+          <SubmitAudioHeading/>
 
           <Stack width= "100%" direction ="row" alignItems="center" spacing={22} sx={{ justifyContent: "center" }} >
             <UploadAudio></UploadAudio>
@@ -21,6 +22,10 @@ export default function Action() {
           </Stack>
 
         </Stack>
+
+      <Link href="/pages/results"> 
+        <span style={{ textDecoration: "underline"}}>Results</span>
+      </Link>
       
     </ThemeProvider>
   );
